Add getBadges to User.
This is intended to allow clients to interpret badges as more than a
bitfield.  We currently return a bse64 string of the redis bitfield.
This method might get parsed in the future to be more user friendly.
Perhaps an array of booleans or array of indexes of current badges.
This all requires more computing on the server's end.