Add score functionallity to User.
Scores were originally intended to have properties such as a timestamp
and type, but this overcomplicated things for no benefit.  So I went
with a simple int that will be the user's score rather than an array of
score objects.  This should allow getting a scoreboard together a lot
easier.