Add User#delete tests
These tests can only test if it exists and is a function as redis-mock
doesn't support SCAN which delete uses.  I've already written a PR for
redis-mock to add SCAN but it most likely won't be added, as it is a
dead repository.  So until then, we can only test that it exists.